requirements.txt
bandit==1.6.2
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
django-extensions==3.0.1
django-filter==2.3.0
django-heroku==0.3.1
django-import-export==2.2.0
django-tagging==0.5.0
django-taggit==1.3.0
django-url-filter==0.3.15
djangorestframework==3.11.0
gitdb==4.0.5
GitPython==3.1.3
gunicorn==20.0.4
Jinja2==2.11.2
Pillow==7.2.0
pipenv==2020.6.2
psycopg2==2.8.5
pylint==2.5.3
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
sqlparse==0.3.1
urllib3==1.24.3
virtualenv==20.0.25
virtualenv-clone==0.5.4
whitenoise==5.1.0

